requirements.txt
google-api-python-client==2.108.0
opencv-python==4.8.*
numpy==1.21.*
pandas==1.3.*
Pillow==9.5.*
scikit-learn==1.0.*
xgboost==1.6.*
shap==0.42.*
matplotlib==3.5.*
seaborn==0.12.*
face-recognition==1.3.0
pytesseract==0.3.10
tensorflow==2.11.*
nltk==3.8.1
python-dotenv==0.21.*
tqdm==4.66.1
textblob==0.17.*
